Support integers of arbitrary bit width

This branch adds support for integers of arbitrary bit width, both in the
`ir::circuit` representation and in the zkif backend.  (The SCALE backend still
requires `u64` everywhere - if we want to keep using it, we'll need a lowering
pass to split wider integers into 64-bit chunks.)  A later branch will add code
for packing mem/fetch ports into wide integers (such as `u96`) so they can be
sorted with fewer gates.

This is an overhaul of the `Num` and `Int` representations to support
arbitrary-sized integers and to fix some issues with the handling of negative
numbers.

`Int` is a vector of `bits`, as before.  It represents a value in the range `0
.. 2^(bits.len())`.

`Num` uses a single `Scalar` to represent a value in the range `0 ..
2^valid_bits` (`valid_bits` is a new field of the `Num` type).  However, the
`Scalar` is bounded to larger range `0 .. 2^real_bits`, where `valid_bits <=
real_bits <= Scalar::CAPACITY`.  The bits in positions `valid_bits ..
real_bits` are essentially scratch space used by various operations.  We need
to know `real_bits` (i.e., the full range of possible `Scalar` values) to
perform the conversion to `Int`, but only the lower `valid_bits` are kept (the
high bits are discarded).

For example, consider an `Add` gate on `u8`s.  Each input is 8 bits, and the
output is also 8 bits.  Specifically, we start with two `Num` inputs, each
having `real_bits == valid_bits == 8`.  The output still has `valid_bits == 8`,
but `real_bits == 9`, because additions like `255 + 1` can overflow past 8
bits.  If we convert this result back to an `Int`, it will first be converted
to the 9-bit number `0b1_0000_0000` (256), then truncated to the 8-bit value
`0b0000_0000` (0), as `valid_bits == 8` indicates that only the lower 8 bits
are meaningful.

If enough gates are chained together, we could end up adding two `Num`s where
`valid_bits == 8` but `real_bits == 127`.  The output would have `real_bits ==
128`, which exceeds `Scalar::CAPACITY`, meaning this addition could overflow.
To avoid overflow, the backend automatically "truncates" the inputs when needed
by converting them to `Int` and back.  In this case, the result of truncation
would be `Num`s with `real_bits == valid_bits == 8`, which can be added safely.

We must also avoid underflow in subtraction and negation.  An 8-bit subtraction
like `1 - 2` might underflow, so instead we compute `1 + 256 - 2`.  Adding
`256` has no effect on the `valid_bits` of the result, but prevents underflow
from occurring.  In this case, the result is `0b1111_1111`, the
two's-complement representation of `-1`.  Negation works similarly: we compute
`-x` as `2^real_bits - x` to avoid underflow.  As an example, `-0_i8` is
computed as `256 - 0 = 256`; `valid_bits` is still `8`, so the logical value of
the result is `0`, as expected.  (Note that the offsets given as `256` for
these examples are based on `real_bits`, not `valid_bits`.  The `Scalar` value
can range up to `2^real_bits`, so adding only `2^valid_bits` is not sufficient
to prevent underflow.)

See merge request fromager/cheesecloth/witness-checker!7

+++ src/back/zkif/mod.rs
@@ -3,5 +3,5 @@ mod representer;
 mod num;
 mod bit_width;
 mod int_ops;
-mod int64;
+mod int;
 mod field;
